Tests: extract helper functions into separate module
