+++ test/fixtures/config/decorators-legacy/babel.config.decorators-legacy.js
